Move ConvertLinksToPlainText to MarkdownLinkConverter and fix whitespace trimming
- Extract ConvertLinksToPlainText and LinkRegex from MarkdownToSpectreConverter into new MarkdownLinkConverter type
- Fix link text whitespace: trim leading/trailing whitespace in regex-based plain text conversion
- Update ConsoleActivityLogger caller to use MarkdownLinkConverter
- Add MarkdownLinkConverterTests with whitespace trimming coverage
